requirements.txt
Django==3.2.7
djangorestframework==3.12.4
drf-yasg==1.20.0
redgreenunittest==0.1.1
coverage==5.5
